requirements.txt
# AURA - Autonomous Utility Recognition Assistant
# Core ML and Computer Vision
ultralytics>=8.0.196
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# Data Processing and Utilities
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Real-time Dashboard and Web Interface
streamlit>=1.25.0
plotly>=5.15.0
dash>=2.13.0
flask>=2.3.0

# Logging and Monitoring
loguru>=0.7.0
tqdm>=4.65.0

# Configuration and Data Handling
PyYAML>=6.0
python-dotenv>=1.0.0
configparser>=5.3.0

# Image Processing and Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Falcon Simulation Integration
requests>=2.31.0
websocket-client>=1.6.0

# Performance and Utilities
psutil>=5.9.0
